fix(body-line-break-punctuation): 句読点と改行のルールを3つに分割
`body-line-break-punctuation`は唐突な改行の抑制を1つのルールに統合していましたが、
読点・句点・行末の3つの観点をORで1つのエラーメッセージにまとめていたため、
リント失敗時にどの観点で違反したのかが分かりにくく、
LLMが原因を読み間違えることが頻発していました。

そこで以下の3つのルールに分割し、
それぞれが専用のエラーメッセージを返すようにします。

- `body-comma-needs-break`: 読点があるのに改行していない
- `body-period-needs-break`: 句点があるのに改行していない
- `body-break-needs-punctuation`: 改行があるのに句読点で終わっていない

論理行の抽出(`extractLines`)や文字集合・閾値・終端正規表現といった内部ロジックは、
`body-punctuation`ディレクトリの共有モジュールにまとめて3ルールから参照します。
行中の句読点判定パーサは句点専用と読点専用に分割し、
各パーサが自分の検査対象ではない種類の句読点を無害文字として消費することで、
読点の前置文字数による閾値判定などの既存挙動を維持しています。

後方互換は不要なため旧ルールと旧テストは削除し、
テストも3ファイルへ観点ごとに振り分けました。

close #110
